+++ packages/plugins/type-distributor/src/headless/stores/dnd/dnd.store.svelte.ts
@@ -105,7 +105,8 @@ class UseDndStore {
 						iedName: targetSIedName,
 						sourceFunction: functionFromSSD,
 						equipmentUuid,
-						equipmentMatches
+						equipmentMatches,
+						functionScopeUuid
 					})
 				)
 			}